+++ app/__init__.py
@@ -3,8 +3,8 @@ from flask import Flask
 
 app = Flask(__name__)
 
-app.config['VERSION'] = '0.43.0'
-app.config['VERSION_DATE'] = '31 August 2026'
+app.config['VERSION'] = '0.44.0'
+app.config['VERSION_DATE'] = '1 September 2026'
 app.config['VERSION_DATE_ISO'] = datetime.strptime(app.config['VERSION_DATE'], '%d %B %Y').strftime('%Y-%m-%d')
 
 @app.context_processor